Autoencoders (AE)/requirements.txt
absl-py==1.4.0
accelerate==0.21.0
aiohttp==3.8.3
aiosignal==1.2.0
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
backports.functools-lru-cache==1.6.5
blinker==1.4
Bottleneck==1.3.5
brotlipy==0.7.0
cachetools==4.2.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==2.0.4
click==8.0.4
cloudpickle==2.2.1
colorama==0.4.6
contourpy==1.0.5
cryptography==39.0.1
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.1
diffusers==0.10.2
entrypoints==0.4
executing==1.2.0
fastjsonschema==2.17.1
filelock==3.12.2
fonttools==4.25.0
frozenlist==1.3.3
fsspec==2023.6.0
ftfy==6.1.1
google-auth==2.6.0
google-auth-oauthlib==0.4.4
grpcio==1.48.2
huggingface-hub==0.16.4
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.2.0
ipykernel==6.15.0
ipython==8.12.0
jedi==0.18.2
Jinja2==3.1.2
jsonschema==4.18.3
jsonschema-specifications==2023.6.1
jupyter_client==7.4.9
jupyter_core==5.3.1
kiwisolver==1.4.4
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.7.1
matplotlib-inline==0.1.6
mkl-fft==1.3.6
mkl-random==1.2.2
mkl-service==2.4.0
mpmath==1.2.1
multidict==6.0.2
munkres==1.1.4
nbformat==5.9.1
nest-asyncio==1.5.6
networkx==2.8.4
numexpr==2.8.4
numpy==1.24.3
oauthlib==3.2.2
opencv-python==4.8.0.74
packaging==23.0
pandas==1.5.3
parso==0.8.3
pickleshare==0.7.5
Pillow==9.4.0
pip==23.1.2
pkgutil_resolve_name==1.3.10
platformdirs==2.5.2
plotly==5.15.0
ply==3.11
prompt-toolkit==3.0.39
protobuf==3.20.3
psutil==5.9.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pygad==3.1.0
Pygments==2.15.1
PyJWT==2.4.0
pyOpenSSL==23.0.0
pyparsing==3.0.9
PyQt5==5.15.7
PyQt5-sip==12.11.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.7
pywin32==305.1
PyYAML==6.0
pyzmq==25.1.0
referencing==0.29.1
regex==2023.6.3
requests==2.29.0
requests-oauthlib==1.3.0
rpds-py==0.8.11
rsa==4.7.2
safetensors==0.3.1
scipy==1.10.1
setuptools==67.8.0
sip==6.6.2
six==1.16.0
stack-data==0.6.2
sympy==1.11.1
tenacity==8.2.2
tensorboard==2.10.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tokenizers==0.13.3
toml==0.10.2
torch==2.0.1
torch-directml==0.2.0.dev230426
torchvision==0.15.1
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.31.0
typing_extensions==4.6.3
urllib3==1.26.16
wcwidth==0.2.6
Werkzeug==2.2.3
wget==3.2
wheel==0.38.4
win-inet-pton==1.1.0
yarl==1.8.1
zipp==3.11.0
